chore: updates docs + tests
