+++ .gitignore
@@ -114,3 +114,4 @@ e2e/test-project/
 e2e/results/
 .claude
 .astro
+.kiro
