+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "fast-brake",
   "version": "0.0.1",
-  "description": "Fast ES feature detection with zero runtime dependencies",
+  "description": "Blazing fast ES feature detection with zero runtime dependencies. Analyzes JavaScript code to detect ECMAScript features and verify compatibility with target ES versions using a revolutionary two-phase detection system.",
   "main": "dist/index.cjs",
   "module": "dist/index.js",
   "type": "module",
@@ -23,18 +23,23 @@
     "commit": "git-cz",
     "commit-msg": "commitlint --edit $1",
     "dev": "bun run --cwd site dev",
-    "lint": "oxlint src test",
-    "lint-fix": "oxlint src test --fix",
+    "lint": "oxlint src tests",
+    "lint-fix": "oxlint src tests --fix",
     "lint-src": "oxlint src",
     "pre-commit": "lint-staged --quiet && bun run test",
     "prepare": "is-ci || husky install",
     "prepublishOnly": "bun run test && bun run build",
     "preview": "bun run --cwd site preview",
     "release": "release-it",
-    "test": "bun test src/*.test.ts test/index.test.ts --timeout 5000",
-    "test:e2e": "bun test test/e2e.test.ts --timeout 5000",
-    "test-e2e": "bun test test/e2e.test.ts --timeout 5000",
-    "benchmark": "bun test/benchmark.ts",
+    "test": "bun test src/*.test.ts tests/index.test.ts --timeout 5000",
+    "test:all": "bun test src/*.test.ts tests/*.test.ts --timeout 5000",
+    "test:unit": "bun test src/*.test.ts --timeout 5000",
+    "test:integration": "bun test tests/index.test.ts --timeout 5000",
+    "test:e2e": "bun test tests/e2e.test.ts --timeout 5000",
+    "test-e2e": "bun test tests/e2e.test.ts --timeout 5000",
+    "test:watch": "bun test --watch src/*.test.ts tests/*.test.ts",
+    "benchmark": "bun tests/benchmark.ts",
+    "benchmark:quick": "bun tests/benchmark.ts --quick",
     "typecheck": "turbo run typecheck-src",
     "typecheck-src": "tsc --noEmit",
     "update": "codependence --update && bun run update-site-deps",
@@ -45,19 +50,25 @@
     "url": "git+https://github.com/yowainwright/fast-brake.git"
   },
   "keywords": [
-    "build",
-    "performance",
-    "caching",
-    "optimization",
-    "dependencies",
-    "build-tools",
-    "typescript",
-    "cli",
     "es",
     "ecmascript",
     "feature-detection",
     "compatibility",
-    "lint",
+    "javascript",
+    "typescript",
+    "ast",
+    "parser",
+    "tokenizer",
+    "babel",
+    "transpiler",
+    "build-tools",
+    "webpack",
+    "rollup",
+    "vite",
+    "eslint",
+    "performance",
+    "fast",
+    "zero-dependencies",
     "es5",
     "es6",
     "es2015",
@@ -67,7 +78,24 @@
     "es2019",
     "es2020",
     "es2021",
-    "es2022"
+    "es2022",
+    "arrow-functions",
+    "template-literals",
+    "async-await",
+    "optional-chaining",
+    "nullish-coalescing",
+    "private-fields",
+    "static-blocks",
+    "bigint",
+    "classes",
+    "destructuring",
+    "spread-operator",
+    "for-of",
+    "generators",
+    "promises",
+    "modules",
+    "import",
+    "export"
   ],
   "author": "Jeff Wainwright <https:jeffry.in> (yowainwright@gmail.com)",
   "license": "MIT",
@@ -113,6 +141,9 @@
   },
   "lint-staged": {
     "src/**/*.ts": [
+      "oxlint --fix"
+    ],
+    "tests/**/*.ts": [
       "oxlint --fix"
     ]
   },
