+++ tests/fixtures/es2015.js
