+++ tests/fixtures/es2017.js
