+++ tests/fixtures/es2022.js
