+++ tests/fixtures/es5.js
